Added 300LA load model
